Chore: Use Hooks for handling state and separation into components
